moves discord service creation to separate class as the entry point is growing
